updated css

+++ practice-for-week-15-react-redux-authenticate-me-long-practice/authenticate-me/frontend/src/components/Cocktails/singleCocktail.js
@@ -9,6 +9,8 @@ const SingleCocktail = () => {
     const { id } = useParams();
     const dispatch = useDispatch();
     const cocktail = useSelector(state => state.cocktail.cocktails[id])
+    const user = useSelector(state => state.session.user)
+
     const history = useHistory()
 
     useEffect(() => {
@@ -22,24 +24,43 @@ const SingleCocktail = () => {
     const [closeForm, openForm] = useState(false);
 
 
-    const handleDelete = async (id) =>{
+    const handleDelete = async (id) => {
         await dispatch(removeCocktail(id))
         history.push('/cocktails')
     }
-    
-    return (
-        <div className="single-cocktail">
-            <h2>{cocktail?.name}</h2>
-            <p>{cocktail?.liquorType}</p>
-            <p>{cocktail?.ingredients}</p>
-            <p>{cocktail?.directions}</p>
-            <p>{cocktail?.imgUrl}</p>
-            <button type='button' onClick={() => handleDelete(cocktail.id)}>Delete</button>
-            <button type='button' onClick={editForm}>Edit</button>
-            {closeForm && (<EditCocktailForm />)}
-            
-        </div>
-    )
+
+    if (cocktail.userId === user.id) {
+        return (
+            <div className="single-cocktail">
+                <div
+                    className="single-cocktail-image"
+                    style={{ backgroundImage: `url('${cocktail?.imgUrl}')` }}
+                ></div>
+                <h2>{cocktail?.name}</h2>
+                <p>{cocktail?.liquorType}</p>
+                <p>{cocktail?.ingredients}</p>
+                <p>{cocktail?.directions}</p>
+                <button type='button' onClick={() => handleDelete(cocktail.id)}>Delete</button>
+                <button type='button' onClick={editForm}>Edit</button>
+                {closeForm && (<EditCocktailForm openForm={openForm} />)}
+            </div>
+        )
+    } else {
+        return (
+            <div className="single-cocktail">
+                <div
+                    className="single-cocktail-image"
+                    style={{ backgroundImage: `url('${cocktail?.imgUrl}')` }}
+                ></div>
+                <div className="cocktail-info">
+                    <h2>{cocktail?.name}</h2>
+                    <p>Liquor Type:{cocktail?.liquorType}</p>
+                    <p>Ingredients:{cocktail?.ingredients}</p>
+                    <p>How to make:{cocktail?.directions}</p>
+                </div>
+            </div>
+        )
+    }
 }
 
 export default SingleCocktail
